Se añadio el metodo de ubicar a los jugadores en los vestuarios

+++ src/model/Price.java
@@ -1,5 +1,5 @@
 package model;
 public interface Price{
 	public double calculateMarketPrice();
-	public int calculateStarLevel();
+	public double calculateStarLevel();
 }
